Better integration with sus.
