+++ config.ru
@@ -1,2 +1,2 @@
-require_relative 'lib/rack/conform/server'
-run Rack::Conform::Server.new
+require_relative 'lib/rack/conform/application'
+run Rack::Conform::Application.new
