+++ lib/rack/conform.rb
@@ -4,4 +4,4 @@
 # Copyright, 2022, by Samuel Williams.
 
 require_relative 'conform/version'
-require_relative 'conform/server'
+require_relative 'conform/application'
